Update and rename Endangered.js to Demonic.js
