+++ Games/types/Mafia/roles/cards/Demonic.js
@@ -1,0 +1,87 @@
+const Card = require("../../Card");
+const { CULT_FACTIONS,MAFIA_FACTIONS } = require("../../const/FactionList");
+
+module.exports = class Demonic extends Card {
+  constructor(role) {
+    super(role);
+
+    if (role.alignment == "Independent") {
+      return;
+    }
+
+    this.listeners = {
+      roleAssigned: function (player) {
+        if (player !== this.player) {
+          return;
+        }
+
+        if ((this.player.hasItem("IsTheTelevangelist") || this.player.role.name == "Televangelist") || CULT_FACTIONS.includes(this.player.faction)) {
+          this.player.queueAlert(
+              `The Cult has Choosen you to perform the Demonic Ritual! Cult dies if no Demonic Players are alive so You must survive!`
+            );
+        }
+        else if (MAFIA_FACTIONS.includes(this.player.faction)) {
+          this.player.queueAlert(
+              `The Cult has Choosen you to perform the Demonic Ritual! Cult dies if no Demonic Players are alive so This could be useful!`
+            );
+        }
+        else if(this.player.faction == "Independent"){
+          this.player.queueAlert(
+              `The Cult has Choosen you to perform the Demonic Ritual! Cult dies if no Demonic Players are alive so This will make things interesting!`
+            );
+        }
+        else{
+          this.player.queueAlert(
+              `The Cult has Choosen you to perform the Demonic Ritual! Cult dies if no Demonic Players are alive so you must die to Stop them!`
+            );
+        }
+        if (this.player.hasItem("IsTheTelevangelist")|| this.player.role.name == "Televangelist") {
+          return;
+        }
+        this.game.queueAlert(
+          `${this.player.role.name} is Endangered! Don't let them all die!`,
+          0,
+          this.game.players.filter((p) => CULT_FACTIONS.includes(p.faction) && p != this.player)
+        );
+      },
+      death: function (player, killer, killType, instant) {
+        var aliveRoles = this.game.players.filter((p) => p.alive && this.game.getRoleTags(p.role.name,p.role.modifier).includes("Demonic"));
+        if (aliveRoles.length > 0) {
+          return;
+        }
+
+        if (this.player.hasItem("IsTheTelevangelist")|| this.player.role.name == "Televangelist") {
+          return;
+        }
+
+          var devotion = this.game.players.filter(
+            (p) => p.alive && p.role.data.DevotionCult
+          );
+          if (devotion.length > 0) {
+            var backUpTarget = devotion.filter(
+              (p) => p.role.data.BackUpConvert
+            );
+            if (backUpTarget.length > 0) {
+              backUpTarget.setRole(
+                `${this.player.role.name}:${this.player.role.modifier}`,
+                this.player.role.data,
+                false,
+                false,
+                false,
+                "No Change"
+              );
+              return;
+            }
+            this.game.events.emit("Devotion", this.player);
+            return;
+          }
+
+        for (let p of this.game.alivePlayers()) {
+          if (CULT_FACTIONS.includes(p.faction)) {
+            p.kill("basic", this.player, instant);
+          }
+        }
+      },
+    };
+  }
+};
